backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-generativeai
passlib[argon2]
python-jose
argon2-cffi
requests
httpx
pytest
psycopg2-binary
